refactor(eslint-rules): inline runtime helpers into base-hook-no-forbidden-runtime
The shared utilities split off in the previous commit were only consumed by
`base-hook-no-forbidden-runtime`:

- `utils/tracked-imports.ts`
- `utils/transitive-reach.ts`
- `utils/module-resolver.ts`

Inline them back into the rule file so `utils/` only contains genuinely
cross-rule helpers (currently `base-hook-detector.ts`, used by both the
signature and runtime rules).

Also drop the `export` keyword from `BASE_SUFFIX` / `UNSTABLE_SUFFIX` in
`base-hook-detector.ts` — both are only consumed by the in-file
`createPairDetector` factory.
